# rpc: method-specific parameters parsing optimisation