--- a/pkg/rpc/client/wsclient_test.go
+++ b/pkg/rpc/client/wsclient_test.go
@@ -187,10 +187,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.BlockFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, 3, filt.Primary)
 			},
 		},
@@ -202,10 +200,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.TxFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, util.Uint160{1, 2, 3, 4, 5}, *filt.Sender)
 				require.Nil(t, filt.Signer)
 			},
@@ -218,10 +214,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.TxFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Nil(t, filt.Sender)
 				require.Equal(t, util.Uint160{0, 42}, *filt.Signer)
 			},
@@ -235,10 +229,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.TxFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, util.Uint160{1, 2, 3, 4, 5}, *filt.Sender)
 				require.Equal(t, util.Uint160{0, 42}, *filt.Signer)
 			},
@@ -251,10 +243,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.NotificationFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, util.Uint160{1, 2, 3, 4, 5}, *filt.Contract)
 				require.Nil(t, filt.Name)
 			},
@@ -267,10 +257,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.NotificationFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, "my_pretty_notification", *filt.Name)
 				require.Nil(t, filt.Contract)
 			},
@@ -284,10 +272,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.NotificationFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, util.Uint160{1, 2, 3, 4, 5}, *filt.Contract)
 				require.Equal(t, "my_pretty_notification", *filt.Name)
 			},
@@ -300,10 +286,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 			},
 			func(t *testing.T, p *request.Params) {
 				param := p.Value(1)
-				raw, ok := param.Value.(json.RawMessage)
-				require.True(t, ok)
 				filt := new(request.ExecutionFilter)
-				require.NoError(t, json.Unmarshal(raw, filt))
+				require.NoError(t, json.Unmarshal(param.RawMessage, filt))
 				require.Equal(t, "FAULT", filt.State)
 			},
 		},
@@ -320,9 +304,8 @@ func TestWSFilteredSubscriptions(t *testing.T) {
 					req := request.In{}
 					err = ws.ReadJSON(&req)
 					require.NoError(t, err)
-					params, err := req.Params()
-					require.NoError(t, err)
-					c.serverCode(t, params)
+					params := request.Params(req.RawParams)
+					c.serverCode(t, &params)
 					err = ws.SetWriteDeadline(time.Now().Add(2 * time.Second))
 					require.NoError(t, err)
 					err = ws.WriteMessage(1, []byte(`{"jsonrpc": "2.0", "id": 1, "result": "0"}`))