--- a/pkg/rpc/request/params.go
+++ b/pkg/rpc/request/params.go
@@ -17,15 +17,6 @@ func (p Params) Value(index int) *Param {
 	return nil
 }
 
-// ValueWithType returns the param struct at the given index if it
-// exists and matches the given type.
-func (p Params) ValueWithType(index int, valType paramType) *Param {
-	if val := p.Value(index); val != nil && val.Type == valType {
-		return val
-	}
-	return nil
-}
-
 func (p Params) String() string {
 	return fmt.Sprintf("%v", []Param(p))
 }