--- a/pkg/rpc/request/types.go
+++ b/pkg/rpc/request/types.go
@@ -53,7 +53,7 @@ type Request struct {
 type In struct {
 	JSONRPC   string          `json:"jsonrpc"`
 	Method    string          `json:"method"`
-	RawParams json.RawMessage `json:"params,omitempty"`
+	RawParams []Param         `json:"params,omitempty"`
 	RawID     json.RawMessage `json:"id,omitempty"`
 }
 
@@ -133,17 +133,4 @@ func NewIn() *In {
 	return &In{
 		JSONRPC: JSONRPCVersion,
 	}
-}
-
-// Params takes a slice of any type and attempts to bind
-// the params to it.
-func (r *In) Params() (*Params, error) {
-	params := Params{}
-
-	err := json.Unmarshal(r.RawParams, &params)
-	if err != nil {
-		return nil, fmt.Errorf("error parsing params: %w", err)
-	}
-
-	return &params, nil
 }