### Problem

Close #2145.

### Solution

Unmarshal parameters to `[]json.RawMessage` and perform further unmarshalling on demand.

TODO:

- [x] Fix failing tests.
- [x] Revisit all usages of `GetInt()`/`AsInt()`, `GetBool()/AsBool()` and `GetString()/AsString()` and adjust wrt reference implementation.
- [x] Benchmarks.