@AnnaShaleva 2021-11-01
I wrote another benchmark that includes not only `request.In` unmarshalling, but also `handleIn` handler (it should be included in benchmark because it also does unmarshalling-related things under the hood). And the optimisation is finally here:
```
name                          old time/op    new time/op    delta
HandleIn/no_extra_params-8      17.2µs ± 5%    15.4µs ± 4%  -10.26%  (p=0.000 n=10+10)
HandleIn/with_extra_params-8    76.4µs ± 9%    44.5µs ±14%  -41.73%  (p=0.000 n=9+10)

name                          old alloc/op   new alloc/op   delta
HandleIn/no_extra_params-8      5.72kB ±16%    3.61kB ±29%  -36.79%  (p=0.000 n=10+10)
HandleIn/with_extra_params-8    70.1kB ±28%    17.0kB ±53%  -75.79%  (p=0.000 n=10+10)

name                          old allocs/op  new allocs/op  delta
HandleIn/no_extra_params-8        47.0 ± 0%      45.0 ± 0%   -4.26%  (p=0.000 n=10+10)
HandleIn/with_extra_params-8       425 ± 0%       217 ± 0%  -48.94%  (p=0.000 n=10+10)
```